Refactor & cleanup
