+++ BetterHands/Configs/MagPalmControls.cs
